moved purchases to home component
